BUGFIX: Do not accumulate multiple EOL markers above the line being edited.
Because removal of previous adornments was moved out of CreateLineVisuals(),
typing text on a line will cause multiple end-of-line markers to be overlaid
on top of each other on the the previous line. The code used to indiscriminately
recreate the adornment of the previous line of an edit as a way to deal with
the situation where pressing enter on an empty line causes the adornment of
the line to be removed but only a layout changed event to be generated for the
next line.

The code now only clears the adornments of the single line above the changed
span of lines, before recreating the adornment. That way the result is correct
regardless of whether the edit operation itself cleared the annotations on that
line.

Refactored the code to make it possible to add a unit test to verify that
we're not overlaying multiple adornments.