Work around missing reformatted line event in multiple runs of lines.
VS2013 and VS2015 Preview under certain circumstances leave out the first line
in runs of lines affected by changes.

For example when pressing enter on an empty line it removes the adornment of
the empty line, but only sends the newly created line below in
e.NewOrReformattedLines.

A single call to OnLayoutChanged can contain multiple runs of affected lines,
and each run may be missing a first line which have been stripped of its
previous adornments.

Added code and test for handling this scenario.

Fixes issue #5.